[MINOR] feat(server,dashboard,coordinator): Report configured metrics of server to coordinator and display to dashboard (#2239)
### What changes were proposed in this pull request?

Report configured metrics of server to coordinator and display to dashboard

### Why are the changes needed?

With this feature, it can be extensible for displaying server state by dashboard.

In the beginning, we only need to display the app number in the dashboard, after a few days, we want to add partition with node in to dashboard, maybe no more days later, we need another value displayed in to dashboard.

So it is not extensible if we add it one by one, this is why I invent this feature, it can be easy to let user config the metrics what they think it is necessary and should be displayed in the dashboard.

### Does this PR introduce _any_ user-facing change?

- new config: rss.server.displayMetricsList

### How was this patch tested?

Test Locally.

<img width="288" alt="image" src="https://github.com/user-attachments/assets/a10fdbe4-08fa-4ccb-8614-80b2447b30f9">

(cherry picked from commit 910823d811c83aee3438647882435aa0efe9e9b1)